requirements.txt
botcity-framework-core>=0.2.7
botcity-maestro-sdk>=0.1.4
# Uncomment to use botcity-utils
#botcity-utils>=0.1.2
# Uncomment to use botcity-crawler-plugin
#botcity-crawler-plugin